Refactor code into packages
The project currently has a single directory structure.

To improve organization, modularity, and maintainability, the code will be refactored into distinct packages. This will group related classes together and improve the structure of the project.

Let's create the command, storage, ui, and task packages and move the related files into these packages.

This separation will lay the foundation for a cleaner, more modular codebase that is easier to navigate and extend.
