+++ src/main/java/abuhurairah/command/AddCommand.java
@@ -1,4 +1,7 @@
-package abuhurairah;
+package abuhurairah.command;
+
+import abuhurairah.task.*;
+import abuhurairah.ui.Ui;
 
 import java.util.ArrayList;
