+++ src/main/java/abuhurairah/storage/Parser.java
@@ -1,4 +1,6 @@
-package abuhurairah;
+package abuhurairah.storage;
+
+import abuhurairah.task.TaskList;
 
 import java.util.Arrays;
