+++ src/main/java/abuhurairah/storage/Storage.java
@@ -1,4 +1,8 @@
-package abuhurairah;
+package abuhurairah.storage;
+
+import abuhurairah.task.TaskList;
+import abuhurairah.task.Task;
+import abuhurairah.ui.Ui;
 
 import java.io.File;
 import java.io.FileNotFoundException;
