+++ src/main/java/abuhurairah/task/Deadline.java
@@ -1,4 +1,4 @@
-package abuhurairah;
+package abuhurairah.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
