+++ src/main/java/abuhurairah/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package abuhurairah;
+package abuhurairah.ui;
 
 import java.io.IOException;
 import java.util.Collections;
